refactor: rename privaticy button
