+++ src/pages/EditEvent.tsx
@@ -1,5 +1,5 @@
 import NavigationButton from "@/components/buttons/navigation-button";
-import PrivacityButton from "@/components/buttons/PrivacityButton/PrivacityButton";
+import PrivacityButton from "@/components/buttons/privacity-button";
 import Tag from "@/components/buttons/tag/tag";
 import Form from "@/components/layout/form";
 import Frame from "@/components/layout/frame";
